Move JsonSerializerContext to S.T.J.Serialization namespace and address doc feedback
